+++ src/StockCopilot/Models/StockDataRetriever.cs
@@ -3,5 +3,5 @@ using StockCopilot.Abstractions.Models;
 
 namespace StockCopilot.Models;
 
-public class StockDataRetriever(string name, (Func<KLine, decimal> Getter, Action<KLine, decimal> Setter) data) : 
-    SelectableItem<(Func<KLine, decimal> Getter, Action<KLine, decimal> Setter)>(name, data);
+public class StockDataRetriever(string name, (Func<KLine, float> Getter, Action<KLine, float> Setter) data) : 
+    SelectableItem<(Func<KLine, float> Getter, Action<KLine, float> Setter)>(name, data);
